Create a top level CompilerLogParser class
