+++ src/components/parser/compilerlog.ts
@@ -1,0 +1,134 @@
+import * as vscode from 'vscode'
+import * as fs from 'fs'
+
+import type { Extension } from '../../main'
+import { convertFilenameEncoding } from '../../utils/utils'
+import { LatexLogParser } from './latexlog'
+
+const latexPattern = /^Output\swritten\son\s(.*)\s\(.*\)\.$/gm
+const latexFatalPattern = /Fatal error occurred, no output PDF file produced!/gm
+
+const latexmkPattern = /^Latexmk:\sapplying\srule/gm
+const latexmkLog = /^Latexmk:\sapplying\srule/
+const latexmkLogLatex = /^Latexmk:\sapplying\srule\s'(pdf|lua|xe)?latex'/
+const latexmkUpToDate = /^Latexmk: All targets \(.*\) are up-to-date/
+
+const texifyPattern = /^running\s(pdf|lua|xe)?latex/gm
+const texifyLog = /^running\s((pdf|lua|xe)?latex|miktex-bibtex)/
+const texifyLogLatex = /^running\s(pdf|lua|xe)?latex/
+
+const DIAGNOSTIC_SEVERITY: { [key: string]: vscode.DiagnosticSeverity } = {
+    'typesetting': vscode.DiagnosticSeverity.Information,
+    'warning': vscode.DiagnosticSeverity.Warning,
+    'error': vscode.DiagnosticSeverity.Error,
+}
+
+export interface LogEntry { type: string, file: string, text: string, line: number }
+
+export class CompilerLogParser {
+    // private readonly extension: Extension
+    private readonly latexLogParser: LatexLogParser
+    isLaTeXmkSkipped: boolean = false
+
+    constructor(extension: Extension) {
+        // this.extension = extension
+        this.latexLogParser = new LatexLogParser(extension)
+    }
+
+    parse(log: string, rootFile?: string) {
+        this.isLaTeXmkSkipped = false
+        // Canonicalize line-endings
+        log = log.replace(/(\r\n)|\r/g, '\n')
+
+        if (log.match(latexmkPattern)) {
+            log = this.trimLaTeXmk(log)
+        } else if (log.match(texifyPattern)) {
+            log = this.trimTexify(log)
+        }
+        if (log.match(latexPattern) || log.match(latexFatalPattern)) {
+            this.latexLogParser.parse(log, rootFile)
+        } else if (this.latexmkSkipped(log)) {
+            this.isLaTeXmkSkipped = true
+        }
+    }
+
+    private trimLaTeXmk(log: string): string {
+        const lines = log.split('\n')
+        let startLine = -1
+        let finalLine = -1
+        for (let index = 0; index < lines.length; index++) {
+            const line = lines[index]
+            let result = line.match(latexmkLogLatex)
+            if (result) {
+                startLine = index
+            }
+            result = line.match(latexmkLog)
+            if (result) {
+                finalLine = index
+            }
+        }
+        if (finalLine <= startLine) {
+            return lines.slice(startLine).join('\n')
+        } else {
+            return lines.slice(startLine, finalLine).join('\n')
+        }
+    }
+
+    private trimTexify(log: string): string {
+        const lines = log.split('\n')
+        let startLine = -1
+        let finalLine = -1
+        for (let index = 0; index < lines.length; index++) {
+            const line = lines[index]
+            let result = line.match(texifyLogLatex)
+            if (result) {
+                startLine = index
+            }
+            result = line.match(texifyLog)
+            if (result) {
+                finalLine = index
+            }
+        }
+        if (finalLine <= startLine) {
+            return lines.slice(startLine).join('\n')
+        } else {
+            return lines.slice(startLine, finalLine).join('\n')
+        }
+    }
+
+    private latexmkSkipped(log: string): boolean {
+        const lines = log.split('\n')
+        if (lines[0].match(latexmkUpToDate)) {
+            this.showCompilerDiagnostics(this.latexLogParser.compilerDiagnostics, this.latexLogParser.buildLog)
+            return true
+        }
+        return false
+    }
+
+    showCompilerDiagnostics(compilerDiagnostics: vscode.DiagnosticCollection, buildLog: LogEntry[]) {
+        compilerDiagnostics.clear()
+        const diagsCollection: { [key: string]: vscode.Diagnostic[] } = {}
+        for (const item of buildLog) {
+            const range = new vscode.Range(new vscode.Position(item.line - 1, 0), new vscode.Position(item.line - 1, 65535))
+            const diag = new vscode.Diagnostic(range, item.text, DIAGNOSTIC_SEVERITY[item.type])
+            diag.source = 'LaTeX'
+            if (diagsCollection[item.file] === undefined) {
+                diagsCollection[item.file] = []
+            }
+            diagsCollection[item.file].push(diag)
+        }
+
+        const configuration = vscode.workspace.getConfiguration('latex-workshop')
+        const convEnc = configuration.get('message.convertFilenameEncoding') as boolean
+        for (const file in diagsCollection) {
+            let file1 = file
+            if (!fs.existsSync(file1) && convEnc) {
+                const f = convertFilenameEncoding(file1)
+                if (f !== undefined) {
+                    file1 = f
+                }
+            }
+            compilerDiagnostics.set(vscode.Uri.file(file1), diagsCollection[file])
+        }
+    }
+}
